+++ src/configuration.ml
@@ -31,6 +31,7 @@ type t = {
   alias_barrier_modules : string list;
   constructor_map : ConstructorMapping.t list;
   error_blacklist : string list;
+  error_filename_blacklist : string list;
   escape_value : string list;
   file_name : string;
   first_class_module_path_blacklist : string list;
@@ -50,6 +51,7 @@ let default (file_name : string) : t = {
   alias_barrier_modules = [];
   constructor_map = [];
   error_blacklist = [];
+  error_filename_blacklist = [];
   escape_value = [];
   file_name;
   first_class_module_path_blacklist = [];
@@ -78,6 +80,9 @@ let is_constructor_renamed (configuration : t) (typ : string) (name : string)
 
 let is_error_in_blacklist (configuration : t) (error_id : string) : bool =
   List.mem error_id configuration.error_blacklist
+
+let is_filename_in_error_blacklist (configuration : t) : bool =
+  List.mem configuration.file_name configuration.error_filename_blacklist
 
 let is_value_to_escape (configuration : t) (name : string) : bool =
   List.mem name configuration.escape_value
@@ -199,6 +204,9 @@ let of_json (file_name : string) (json : Yojson.Basic.t) : t =
         | "error_blacklist" ->
           let entry = get_string_list "error_blacklist" entry in
           {configuration with error_blacklist = entry}
+        | "error_filename_blacklist" ->
+          let entry = get_string_list "error_filename_blacklist" entry in
+          {configuration with error_filename_blacklist = entry}
         | "escape_value" ->
           let entry = get_string_list "escape_value" entry in
           {configuration with escape_value = entry}